+++ source/controllers/application_controller.py
@@ -11,11 +11,13 @@ Autore:
 Marco Cantù
 
 Versione:
-0.2.0
+0.5.0
 ==========================================================
 """
 
+from source.controllers.photo_controller import PhotoController
 from source.controllers.project_controller import ProjectController
+from source.services.photo.photo_manager import PhotoManager
 from source.services.project.project_manager import ProjectManager
 
 
@@ -23,20 +25,36 @@ class ApplicationController:
     """
     Controller principale dell'applicazione.
 
-    Espone controller e servizi condivisi
-    a tutta l'applicazione.
+    Istanzia i servizi condivisi e i controller
+    dell'intera applicazione.
     """
 
     def __init__(self) -> None:
 
-        # Controller
-        self.project_controller = ProjectController()
-
+        #
         # Services
+        #
+
         self.project_manager = ProjectManager()
 
+        self.photo_manager = PhotoManager(
+            self.project_manager
+        )
+
+        #
+        # Controllers
+        #
+
+        self.project_controller = ProjectController(
+            self.project_manager
+        )
+
+        self.photo_controller = PhotoController(
+            self.photo_manager
+        )
+
     # ---------------------------------------------------------
-    # Controller
+    # Controllers
     # ---------------------------------------------------------
 
     def get_project_controller(self) -> ProjectController:
@@ -46,11 +64,31 @@ class ApplicationController:
         return self.project_controller
 
     # ---------------------------------------------------------
+
+    def get_photo_controller(self) -> PhotoController:
+        """
+        Restituisce il PhotoController.
+        """
+        return self.photo_controller
+
+    # ---------------------------------------------------------
     # Services
     # ---------------------------------------------------------
+    #
+    # Utilizzati internamente dall'applicazione.
+    # La GUI deve accedere esclusivamente ai Controller.
+    #
 
     def get_project_manager(self) -> ProjectManager:
         """
         Restituisce il ProjectManager.
         """
         return self.project_manager
+
+    # ---------------------------------------------------------
+
+    def get_photo_manager(self) -> PhotoManager:
+        """
+        Restituisce il PhotoManager.
+        """
+        return self.photo_manager